Do not remove transfer caches when the app crashed

+++ src/MEGASync/MegaApplication.cpp
@@ -1055,9 +1055,9 @@ void MegaApplication::initialize()
         {
             di.next();
             const QFileInfo& fi = di.fileInfo();
-            if (fi.fileName().endsWith(QString::fromAscii(".db"))
+            if (!fi.fileName().contains(QString::fromUtf8("transfers_")) && (fi.fileName().endsWith(QString::fromAscii(".db"))
                     || fi.fileName().endsWith(QString::fromAscii(".db-wal"))
-                    || fi.fileName().endsWith(QString::fromAscii(".db-shm")))
+                    || fi.fileName().endsWith(QString::fromAscii(".db-shm"))))
             {
                 QFile::remove(di.filePath());
             }
